Added CallExecution* to network/result.rs
